# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyPetly API Documentation",
        "description": "This is the API documentation for MyPetly.",
        "contact": {
            "email": "support@mypetly.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/collars": {
            "get": {
                "tags": [
                    "Collars"
                ],
                "summary": "Get all collars",
                "description": "Retrieve a list of all collars",
                "operationId": "33f504b31f250406aa38de5237d9f985",
                "responses": {
                    "200": {
                        "description": "List of collars",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nfc_id": {
                                                "type": "string",
                                                "example": "NFC12345"
                                            },
                                            "qr_code_url": {
                                                "type": "string",
                                                "example": "http://example.com/qrcode/1.png"
                                            },
                                            "animal_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-01-25T12:34:56Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-01-25T12:34:56Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detailed error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Collars"
                ],
                "summary": "Create a new collar",
                "description": "Add a new collar to the system",
                "operationId": "b73282a3f9787c41b693e1399ec379c8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nfc_id",
                                    "animal_id"
                                ],
                                "properties": {
                                    "nfc_id": {
                                        "type": "string",
                                        "example": "NFC12345"
                                    },
                                    "animal_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Collar created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nfc_id": {
                                            "type": "string",
                                            "example": "NFC12345"
                                        },
                                        "qr_code_url": {
                                            "type": "string",
                                            "example": "http://example.com/qrcode/1.png"
                                        },
                                        "animal_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-01-25T12:34:56Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-01-25T12:34:56Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "nfc_id": [
                                                    "The nfc_id field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/collars/{id}": {
            "get": {
                "tags": [
                    "Collars"
                ],
                "summary": "Get collar details",
                "description": "Retrieve the details of a specific collar by ID",
                "operationId": "705a8b29096716092441d2917e9c0ac9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the collar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Collar details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nfc_id": {
                                            "type": "string",
                                            "example": "NFC12345"
                                        },
                                        "qr_code_url": {
                                            "type": "string",
                                            "example": "http://example.com/qrcode/1.png"
                                        },
                                        "animal_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-01-25T12:34:56Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-01-25T12:34:56Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Collar not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Collar not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Collars"
                ],
                "summary": "Update a collar",
                "description": "Update the details of an existing collar",
                "operationId": "3c85c200d23edbab5f343f80c09f79ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the collar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nfc_id": {
                                        "type": "string",
                                        "example": "NFC12345"
                                    },
                                    "qr_code_url": {
                                        "type": "string",
                                        "example": "http://example.com/qrcode/1.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Collar updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nfc_id": {
                                            "type": "string",
                                            "example": "NFC12345"
                                        },
                                        "qr_code_url": {
                                            "type": "string",
                                            "example": "http://example.com/qrcode/1.png"
                                        },
                                        "animal_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-01-25T12:34:56Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-01-25T12:34:56Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Collar not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Collar not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Collars"
                ],
                "summary": "Delete a collar",
                "description": "Remove a collar from the system",
                "operationId": "f963f6caa43ad12a3765527937e1f2f9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the collar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Collar deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Collar deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Collar not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Collar not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/collars/{collarId}/assign/{animalId}": {
            "post": {
                "tags": [
                    "Collars"
                ],
                "summary": "Assign a collar to an animal",
                "description": "Link a collar to a specific animal",
                "operationId": "0ccee44831e0cc3143f32c4675906449",
                "parameters": [
                    {
                        "name": "collarId",
                        "in": "path",
                        "description": "ID of the collar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "animalId",
                        "in": "path",
                        "description": "ID of the animal",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Collar assigned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Collar assigned successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Collar or animal not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Collar or animal not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}